simpledokcerapp/myapp/requirements.txt
 
Flask  
pymongo
python-dotenv


